Adds sloppy implementation of multiple camera views at the same time
each camera position is tagged with its editorId, and gets sent out each frame. needed to remove async to do the rendering from each camera. relies on clients filtering out messages not intended for them - but in the future it should just specifically send to the sockets that need it.